prepare clusterintegrationtest pkg for multinode backup tests
