+++ adapters/repos/db/clusterintegrationtest/backup_coordinator_integration_test.go
@@ -1,0 +1,10 @@
+//go:build integrationTest
+// +build integrationTest
+
+package clusterintegrationtest
+
+import (
+	"testing"
+)
+
+func TestDistributedBackups(t *testing.T) {}
