+++ adapters/repos/db/clusterintegrationtest/fakes_for_test.go
@@ -1,0 +1,210 @@
+//go:build integrationTest
+// +build integrationTest
+
+package clusterintegrationtest
+
+import (
+	"context"
+	"net/http"
+	"net/http/httptest"
+	"net/url"
+	"path"
+
+	"github.com/semi-technologies/weaviate/adapters/clients"
+	"github.com/semi-technologies/weaviate/adapters/handlers/rest/clusterapi"
+	"github.com/semi-technologies/weaviate/adapters/repos/db"
+	"github.com/semi-technologies/weaviate/entities/backup"
+	"github.com/semi-technologies/weaviate/entities/models"
+	"github.com/semi-technologies/weaviate/entities/modulecapabilities"
+	"github.com/semi-technologies/weaviate/entities/schema"
+	ubak "github.com/semi-technologies/weaviate/usecases/backup"
+	"github.com/semi-technologies/weaviate/usecases/sharding"
+	"github.com/sirupsen/logrus/hooks/test"
+)
+
+type node struct {
+	name             string
+	shardingState    *sharding.State
+	repo             *db.DB
+	schemaManager    *fakeSchemaManager
+	backupManager    *ubak.Manager
+	coordinator      *ubak.Coordinator
+	clusterAPIServer *httptest.Server
+	migrator         *db.Migrator
+	hostname         string
+}
+
+func (n *node) init(dirName string, shardStateRaw []byte,
+	allNodes *[]*node,
+) {
+	localDir := path.Join(dirName, n.name)
+	logger, _ := test.NewNullLogger()
+
+	nodeResolver := &nodeResolver{
+		nodes: allNodes,
+		local: n.name,
+	}
+
+	shardState, err := sharding.StateFromJSON(shardStateRaw, nodeResolver)
+	if err != nil {
+		panic(err)
+	}
+
+	client := clients.NewRemoteIndex(&http.Client{})
+	n.repo = db.New(logger, db.Config{
+		FlushIdleAfter:            60,
+		RootPath:                  localDir,
+		QueryMaximumResults:       10000,
+		MaxImportGoroutinesFactor: 1,
+	}, client, nodeResolver, nil)
+	n.schemaManager = &fakeSchemaManager{
+		shardState: shardState,
+		schema:     schema.Schema{Objects: &models.Schema{}},
+	}
+
+	n.repo.SetSchemaGetter(n.schemaManager)
+	err = n.repo.WaitForStartup(context.Background())
+	if err != nil {
+		panic(err)
+	}
+
+	backendProvider := &fakeBackupBackendProvider{}
+	n.backupManager = ubak.NewManager(logger, &fakeAuthorizer{}, n.schemaManager, n.repo, backendProvider)
+
+	backupClient := clients.NewClusterBackups(&http.Client{})
+	n.coordinator = ubak.NewCoordinator(&fakeObjectStore{}, n.repo, backupClient, logger, nodeResolver)
+
+	n.migrator = db.NewMigrator(n.repo, logger)
+
+	indices := clusterapi.NewIndices(sharding.NewRemoteIndexIncoming(n.repo))
+	mux := http.NewServeMux()
+	mux.Handle("/indices/", indices.Indices())
+
+	backups := clusterapi.NewBackups(n.backupManager)
+	mux.Handle("/backups/can-commit", backups.CanCommit())
+	mux.Handle("/backups/commit", backups.Commit())
+	mux.Handle("/backups/abort", backups.Abort())
+	mux.Handle("/backups/status", backups.Status())
+
+	srv := httptest.NewServer(mux)
+	u, err := url.Parse(srv.URL)
+	if err != nil {
+		panic(err)
+	}
+	n.hostname = u.Host
+}
+
+type fakeNodes struct {
+	nodes []string
+}
+
+func (f fakeNodes) AllNames() []string {
+	return f.nodes
+}
+
+func (f fakeNodes) LocalName() string {
+	return f.nodes[0]
+}
+
+type fakeSchemaManager struct {
+	schema     schema.Schema
+	shardState *sharding.State
+}
+
+func (f *fakeSchemaManager) GetSchemaSkipAuth() schema.Schema {
+	return f.schema
+}
+
+func (f *fakeSchemaManager) ShardingState(class string) *sharding.State {
+	return f.shardState
+}
+
+func (f *fakeSchemaManager) RestoreClass(ctx context.Context, d *backup.ClassDescriptor) error {
+	return nil
+}
+
+type nodeResolver struct {
+	nodes *[]*node
+	local string
+}
+
+func (r nodeResolver) AllNames() []string {
+	panic("node resolving not implemented yet")
+}
+
+func (r nodeResolver) LocalName() string {
+	return r.local
+}
+
+func (r nodeResolver) NodeHostname(nodeName string) (string, bool) {
+	for _, node := range *r.nodes {
+		if node.name == nodeName {
+			return node.hostname, true
+		}
+	}
+
+	return "", false
+}
+
+type fakeBackupBackendProvider struct{}
+
+func (f *fakeBackupBackendProvider) BackupBackend(backend string) (modulecapabilities.BackupBackend, error) {
+	return &fakeBackupBackend{}, nil
+}
+
+type fakeBackupBackend struct {
+}
+
+func (f *fakeBackupBackend) HomeDir(backupID string) string {
+	return ""
+}
+
+func (f *fakeBackupBackend) GetObject(ctx context.Context, backupID, key string) ([]byte, error) {
+	return nil, nil
+}
+
+func (f *fakeBackupBackend) WriteToFile(ctx context.Context, backupID, key, destPath string) error {
+	return nil
+}
+
+func (f *fakeBackupBackend) SourceDataPath() string {
+	return ""
+}
+
+func (f *fakeBackupBackend) PutFile(ctx context.Context, backupID, key, srcPath string) error {
+	return nil
+}
+
+func (f *fakeBackupBackend) PutObject(ctx context.Context, backupID, key string, byes []byte) error {
+	return nil
+}
+
+func (f *fakeBackupBackend) Initialize(ctx context.Context, backupID string) error {
+	return nil
+}
+
+type fakeObjectStore struct {
+	modulecapabilities.BackupBackend
+}
+
+func (f *fakeObjectStore) Meta(ctx context.Context, backupID string) (*backup.BackupDescriptor, error) {
+	return nil, nil
+}
+
+func (f *fakeObjectStore) PutMeta(ctx context.Context, desc *backup.BackupDescriptor) error {
+	return nil
+}
+
+func (f *fakeObjectStore) GlobalMeta(ctx context.Context, backupID string) (*backup.DistributedBackupDescriptor, error) {
+	return nil, nil
+}
+
+func (f *fakeObjectStore) PutGlobalMeta(ctx context.Context, desc *backup.DistributedBackupDescriptor) error {
+	return nil
+}
+
+type fakeAuthorizer struct{}
+
+func (f *fakeAuthorizer) Authorize(_ *models.Principal, _, _ string) error {
+	return nil
+}
